fix narrowing error, make format-fix

+++ extension/parquet/include/cast_column_reader.hpp
@@ -31,7 +31,7 @@ public:
 	void Skip(idx_t num_values) override;
 	idx_t GroupRowsAvailable() override;
 
-	size_t TotalCompressedSize() override {
+	uint64_t TotalCompressedSize() override {
 		return child_reader->TotalCompressedSize();
 	}
